Adding basic cap3 support
