+++ lib/application_seeds/capistrano/tasks/application_seeds.task
@@ -1,0 +1,18 @@
+namespace :deploy do
+  task :application_seeds do
+    if fetch(:rails_env) == 'production'
+      raise "You cannot run this task in the production environment"
+    end
+
+    within fetch(:latest_release_directory)
+      with rails_env: fetch(:rails_env) do
+        if fetch(:dataset) == ""
+          execute :rake, 'db:seed'
+        else
+          execute :rake, "application_seeds:load\[#{dataset}\]"
+        end
+      end
+    end
+  end
+end
+
